# api-specification.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Social Distribution API Specification",
    "version": "1.0"
  },
  "servers": [
    { "url": "https://social-distribution-w23-t17.herokuapp.com/authors" }
  ],
...
  "paths": {
...
    "/{author_id}/posts/{post_id}/comments/": {
      "parameters": [
        {
          "name": "author_id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "example": "http://localhost/authors/423b8c9a-06dc-47ee-9461-997dacf5f162"
          }
        },
        {
          "name": "post_id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "example": "http://localhost/authors/b797fc58-e267-4c5e-b2c8-2c9838f48ebf/posts/a7df8d91-b874-4778-9eaa-ba0f008621dc"
          }
        }
      ],
      "get": {
        "tags": ["comments"],
        "summary": "Get comments for a post",
        "operationId": "getPostComments",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "type": {
                      "type": "string"
                    },
                    "page": {
                      "type": "integer"
                    },
                    "size": {
                      "type": "integer"
                    },
                    "post": {
                      "type": "string",
                      "example": "http://127.0.0.1:5454/authors/9de17f29c12e8f97bcbbd34cc908f1baba40658e/posts/764efa883dda1e11db47671c4a3bbd9e"
                    },
                    "id": {
                      "type": "string",
                      "example": "http://127.0.0.1:5454/authors/9de17f29c12e8f97bcbbd34cc908f1baba40658e/posts/764efa883dda1e11db47671c4a3bbd9e/comments"
                    },
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Comment"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["comments"],
        "summary": "Create comment for a post",
        "operationId": "createPostComment",
        "requestBody": {
          "description": "Create Comment for a post",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Comment"
              },
              "commentAuthorId": {
                "type": "string",
                "example": "http://127.0.0.1:5454/authors/9de17f29c12e8f97bcbbd34cc908f1baba40658e"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Comment"
                }
              }
            }
          },
          "400": {
            "description": "Badly formed post object"
          },
          "404": {
            "description": "Post or Author doesn't exist"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Author": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "default": "author"
          },
          "id": {
            "type": "string",
            "example": "http://localhost/authors/"
          },
          "displayName": {
            "type": "string",
            "example": "Joe"
          },
          "url": {
            "type": "string",
            "example": "http://localhost/authors/4d53bc84-0caa-4fd3-b2b4-fd762c2d0786"
          },
          "github": {
            "type": "string",
            "example": "http://localhost/authors/4d53bc84-0caa-4fd3-b2b4-fd762c2d0786"
          },
          "profileImage": {
            "type": "string",
            "example": "http://google.com/funnyanimalpic"
          }
        }
      },
...
      "Comment": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "default": "comment"
          },
          "id": {
            "type": "string",
            "example": "http://localhost/authors/11f6bdd5-a60b-4ca0-a395-1fba77f1ede8/posts/b2920040-9c7f-4ad7-bdc2-2ecb61a94217/comments/317bb454-6b52-457a-a97b-cf7b4e24e42f"
          },
          "author": {
            "type": "object",
            "$ref": "#/components/schemas/Author"
          },
          "comment": {
            "type": "string",
            "description": "content of the comment"
          },
          "contentType": {
            "type": "string",
            "example": "text/markdown | text/plain"
          },
          "published": {
            "type": "string",
            "example": "2023-03-09 00:36:00.795000+00:00"
          }
        }
      },
...
    }
  }
}